requirements.txt
Flask==2.3.2
google-cloud-storage==2.10.0
gunicorn==21.2.0